🚨 useRef propsType을 지정할 수 없어 컴포넌트 병합
