+++ client/src/components/lost/LostPostMap.js
@@ -1,9 +1,23 @@
 import React, { useEffect, useState, useRef } from 'react';
-import { Map, MapMarker } from 'react-kakao-maps-sdk';
+import {
+  Map,
+  MapMarker,
+  Circle,
+  Polyline,
+  CustomOverlayMap,
+} from 'react-kakao-maps-sdk';
 import PropTypes from 'prop-types';
 import FindPlaceName from './FindPlaceName';
-import DrawCircle from './DrawCircle';
 
+function DistanceInfo({ distance }) {
+  const totalRadius =
+    distance >= 1000 ? `${(distance / 1000).toFixed(1)}km` : `${distance}m`;
+  return (
+    <ul className="relative bg-white top-1 left-1 p-2 m-0 text-xs border-1 border-solid shadow rounded">
+      <li>총반경 {totalRadius}</li>
+    </ul>
+  );
+}
 function FindLocation({ setAddress, setAddressName, setRadius }) {
   const [position, setPosition] = useState(false);
   const [state, setState] = useState({
@@ -101,7 +115,34 @@ function FindLocation({ setAddress, setAddressName, setRadius }) {
         onMouseMove={handleMouseMove}
       >
         {position && (
-          <DrawCircle position={position} drawingLineRef={drawingLineRef} />
+          <>
+            <Circle
+              center={position.center}
+              radius={position.radius}
+              strokeWeight={1} // 선의 두께입니다
+              strokeColor="#00a0e9" // 선의 색깔입니다
+              strokeOpacity={0.1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
+              strokeStyle="solid" // 선의 스타일입니다
+              fillColor="#00a0e9" // 채우기 색깔입니다
+              fillOpacity={0.2} // 채우기 불투명도입니다
+            />
+            <Polyline
+              path={[position.center, position.mousePosition]}
+              ref={drawingLineRef}
+              strokeWeight={3} // 선의 두께 입니다
+              strokeColor="#00a0e9" // 선의 색깔입니다
+              strokeOpacity={1} // 선의 불투명도입니다 0에서 1 사이값이며 0에 가까울수록 투명합니다
+              strokeStyle="solid" // 선의 스타일입니다
+            />
+            <CustomOverlayMap
+              position={position.mousePosition}
+              xAnchor={0}
+              yAnchor={0}
+              zIndex={1}
+            >
+              <DistanceInfo distance={Math.floor(position.radius)} />
+            </CustomOverlayMap>
+          </>
         )}
 
         {!state.isLoading && (
@@ -138,6 +179,10 @@ FindLocation.propTypes = {
   setAddress: PropTypes.func.isRequired,
   setAddressName: PropTypes.func.isRequired,
   setRadius: PropTypes.func.isRequired,
+};
+
+DistanceInfo.propTypes = {
+  distance: PropTypes.number.isRequired,
 };
 
 export default FindLocation;
